+++ CHANGES.md
@@ -1,3 +1,10 @@
+1.8.4
+=====
+
+* [#86](https://github.com/serokell/log-warper/issues/86):
+  Add lens for changing properties of the particular logger.
+* Ungrade `universum` to the `1.0.2`.
+
 1.8.3
 =====